requirements.txt
datapackage==1.15.2
matplotlib==3.7.1
numpy==1.25.0
numpy==1.24.3
pandas==2.0.2
pandas==2.0.1
Pillow==9.5.0
Pillow==10.0.0
